Split history report loading and cache
