+++ apps/server/vibesensor/use_cases/history/__init__.py
@@ -3,7 +3,9 @@
 Sub-modules
 -----------
 - :mod:`~vibesensor.use_cases.history.runs` — run query and delete services.
-- :mod:`~vibesensor.use_cases.history.reports` — PDF report generation service.
+- :mod:`~vibesensor.use_cases.history.report_loader` — persisted report loading and shaping.
+- :mod:`~vibesensor.use_cases.history.report_cache` — in-memory PDF cache coordination.
+- :mod:`~vibesensor.use_cases.history.reports` — thin PDF report coordinator service.
 - :mod:`~vibesensor.use_cases.history.exports` — CSV/ZIP export service.
 """
