+++ apps/server/vibesensor/use_cases/history/reports.py
@@ -1,4 +1,4 @@
-"""PDF report loading, language resolution, cache, and coordination.
+"""PDF report build coordination for persisted history runs.
 
 Framework-agnostic: raises domain exceptions from ``vibesensor.shared.exceptions``
 rather than HTTP-specific exceptions.  The routes layer translates domain
@@ -7,27 +7,13 @@ exceptions to HTTP status codes.
 
 from __future__ import annotations
 
-import asyncio
-import json
-import logging
-from collections import OrderedDict
 from collections.abc import Callable
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING
 
-from vibesensor.domain import CarSnapshot
-from vibesensor.shared.boundaries.summary_warning import summary_warning_payloads
-from vibesensor.shared.exceptions import AnalysisNotReadyError, ProcessingError
 from vibesensor.shared.ports import RunPersistence, SettingsReader
-from vibesensor.shared.run_context import add_current_context_warnings, current_car_snapshot_token
-from vibesensor.shared.types.json_types import JsonObject, JsonValue, is_json_object
-from vibesensor.use_cases.history.helpers import (
-    HistoryRecord,
-    async_require_run,
-    require_analysis_ready,
-    resolve_run_language,
-    safe_filename,
-)
+from vibesensor.use_cases.history.report_cache import HistoryReportPdfCache
+from vibesensor.use_cases.history.report_loader import HistoryReportRequestLoader
 
 if TYPE_CHECKING:
     from vibesensor.domain import TestRun
@@ -35,11 +21,6 @@ if TYPE_CHECKING:
 #: Callable that turns a persisted analysis dict into PDF bytes.
 #: Signature: ``(analysis_summary, test_run) -> bytes``.
 PdfRendererFn = Callable[[dict, "TestRun | None"], bytes]
-
-LOGGER = logging.getLogger(__name__)
-
-REPORT_PDF_CACHE_MAX_ENTRIES = 16
-ReportPdfCacheKey = tuple[str, str, str | None, int, str, str]
 
 
 @dataclass(frozen=True)
@@ -50,24 +31,13 @@ class HistoryReportPdf:
     filename: str
 
 
-@dataclass(frozen=True)
-class HistoryReportRequest:
-    """Resolved persisted report context ready for PDF generation."""
-
-    cache_key: ReportPdfCacheKey
-    filename: str
-    analysis_summary: JsonObject
-    domain_test_run: TestRun | None = None
-
-
 class HistoryReportService:
-    """Load persisted report data and coordinate cached PDF generation."""
+    """Coordinate cached PDF generation for persisted history reports."""
 
     __slots__ = (
-        "_history_db",
+        "_loader",
         "_pdf_cache",
         "_pdf_renderer",
-        "_settings_store",
     )
 
     def __init__(
@@ -77,13 +47,12 @@ class HistoryReportService:
         *,
         pdf_renderer: PdfRendererFn,
     ) -> None:
-        self._history_db = history_db
+        self._loader = HistoryReportRequestLoader(history_db, settings_store)
         self._pdf_cache = HistoryReportPdfCache()
         self._pdf_renderer = pdf_renderer
-        self._settings_store = settings_store
 
     async def build_pdf(self, run_id: str, requested_lang: str | None) -> HistoryReportPdf:
-        request = await self.load_report_request(run_id, requested_lang)
+        request = await self._loader.load_report_request(run_id, requested_lang)
         cached_pdf = self._pdf_cache.get(request.cache_key)
         if cached_pdf is not None:
             return HistoryReportPdf(content=cached_pdf, filename=request.filename)
@@ -97,133 +66,3 @@ class HistoryReportService:
             run_id=run_id,
         )
         return HistoryReportPdf(content=pdf, filename=request.filename)
-
-    async def load_report_request(
-        self,
-        run_id: str,
-        requested_lang: str | None,
-    ) -> HistoryReportRequest:
-        run = await async_require_run(self._history_db, run_id)
-        analysis = require_analysis_ready(run)
-        if not isinstance(analysis, dict) or not isinstance(analysis.get("findings"), list):
-            raise AnalysisNotReadyError(
-                "Report data unavailable for this run. Re-analyze to regenerate the PDF."
-            )
-
-        requested_lang = self._analysis_language(run, requested_lang)
-        current_active_car_snapshot = (
-            self._settings_store.active_car_snapshot() if self._settings_store is not None else None
-        )
-        warnings = add_current_context_warnings(
-            analysis.get("warnings"),
-            metadata=analysis.get("metadata"),
-            current_active_car_snapshot=current_active_car_snapshot,
-        )
-        analysis_summary = dict(analysis)
-        analysis_summary["warnings"] = cast(JsonValue, summary_warning_payloads(warnings))
-        cache_key = self._report_pdf_cache_key(
-            run,
-            run_id,
-            self._report_pdf_cache_lang(run, requested_lang),
-            current_active_car_snapshot=current_active_car_snapshot,
-        )
-        return HistoryReportRequest(
-            cache_key=cache_key,
-            filename=f"{safe_filename(run_id)}_report.pdf",
-            analysis_summary=analysis_summary,
-        )
-
-    @staticmethod
-    def _metadata_cache_token(metadata: object) -> str:
-        if not is_json_object(metadata):
-            return "{}"
-        return json.dumps(metadata, sort_keys=True, default=str, ensure_ascii=False)
-
-    def _report_pdf_cache_key(
-        self,
-        run: HistoryRecord,
-        run_id: str,
-        requested_lang: str,
-        *,
-        current_active_car_snapshot: CarSnapshot | None,
-    ) -> ReportPdfCacheKey:
-        return (
-            run_id,
-            requested_lang,
-            str(run["analysis_completed_at"]) if "analysis_completed_at" in run else None,
-            int(run.get("sample_count", 0)),
-            self._metadata_cache_token(run.get("metadata", {})),
-            current_car_snapshot_token(current_active_car_snapshot),
-        )
-
-    @staticmethod
-    def _report_pdf_cache_lang(run: HistoryRecord, requested_lang: str) -> str:
-        analysis = run.get("analysis")
-        if is_json_object(analysis):
-            persisted_lang = str(analysis.get("lang") or "").strip().lower()
-            if persisted_lang:
-                return persisted_lang
-        return requested_lang
-
-    @staticmethod
-    def _analysis_language(run: HistoryRecord, requested: str | None) -> str:
-        return resolve_run_language(run, requested)
-
-
-class HistoryReportPdfCache:
-    """LRU PDF cache plus per-key build coordination."""
-
-    __slots__ = ("_entries", "_locks")
-
-    def __init__(self) -> None:
-        self._entries: OrderedDict[ReportPdfCacheKey, bytes] = OrderedDict()
-        self._locks: dict[ReportPdfCacheKey, asyncio.Lock] = {}
-
-    def get(self, cache_key: ReportPdfCacheKey) -> bytes | None:
-        cached_pdf = self._entries.get(cache_key)
-        if cached_pdf is None:
-            return None
-        self._entries.move_to_end(cache_key)
-        return cached_pdf
-
-    async def get_or_build(
-        self,
-        cache_key: ReportPdfCacheKey,
-        build_pdf: Callable[[], bytes],
-        *,
-        run_id: str,
-    ) -> bytes:
-        build_lock = self._locks.setdefault(cache_key, asyncio.Lock())
-        async with build_lock:
-            cached_pdf = self.get(cache_key)
-            if cached_pdf is not None:
-                return cached_pdf
-            try:
-                pdf = await asyncio.to_thread(build_pdf)
-            except Exception as exc:
-                LOGGER.warning("PDF generation failed for run %s", run_id, exc_info=True)
-                self._prune_stale_locks()
-                raise ProcessingError(
-                    "PDF generation failed due to an internal error."
-                    " Please try again or re-analyze this run."
-                ) from exc
-            self._put(cache_key, pdf)
-            return pdf
-
-    def _put(self, cache_key: ReportPdfCacheKey, pdf: bytes) -> None:
-        self._entries[cache_key] = pdf
-        self._entries.move_to_end(cache_key)
-        while len(self._entries) > REPORT_PDF_CACHE_MAX_ENTRIES:
-            evicted_key, _ = self._entries.popitem(last=False)
-            self._locks.pop(evicted_key, None)
-        self._prune_stale_locks()
-
-    def _prune_stale_locks(self) -> None:
-        if len(self._locks) > REPORT_PDF_CACHE_MAX_ENTRIES * 2:
-            stale_keys = [
-                key
-                for key, lock in self._locks.items()
-                if key not in self._entries and not lock.locked()
-            ]
-            for key in stale_keys:
-                self._locks.pop(key, None)
