+++ docs/ai/repo-map.md
@@ -38,7 +38,7 @@ Scope: detailed file layout, entry points, package structure, and module ownersh
 - `infra/config/`: runtime settings store (single `SettingsStore` owns both analysis and device settings) used by runtime wiring and exposed to use-cases through focused shared read ports.
 - `infra/workers/`: worker-pool infrastructure.
 - `use_cases/diagnostics/`: post-stop diagnostics pipeline. See `docs/analysis_pipeline.md` for the full data flow, pipeline steps, and module map. `run_data_preparation.py` owns `PreparedRunData` plus run timing/speed/phase/sensor preparation, `_summary_steps.py` owns the summary step builders (`build_findings_bundle`, `build_sensor_bundle`, `build_run_suitability_bundle`), `_summary_result.py` owns `AnalysisResult` plus final `TestRun` / `DiagnosticCase` / persisted-summary assembly, `findings.py` owns the top-level finding orchestration/finalization entrypoints, `_peak_findings.py` owns persistent-peak support (`PeakFindingAnalyzer`, phase filtering, duplicate suppression), `_reference_findings.py` owns reference-gap and engine/wheel/sample-rate sufficiency helpers, `peak_binning.py` owns persistent-peak accumulation/scoring, `peak_table.py` owns internal peak-table row ranking/statistics, `signal_aggregation.py` owns speed/location summary aggregation, `math_utils.py` owns generic statistics/correlation helpers, `speed_profile_helpers.py` owns speed-profile assembly helpers, `order_analysis.py` now owns the core order-matching/scoring/assembly primitives, `order_pipeline.py` owns `OrderAnalysisSession` plus the `_build_order_findings()`/multi-location orchestration layer above those primitives, and `order_heuristics.py` owns the pure diffuse-excitation, localization-override, and engine-alias heuristics; `top_cause_selection.py` delegates classification and ranking to the domain `Finding`; `location_analysis.py` owns the location-analysis pipeline and `LocationAnalysisResult` typed return; `plots.py` now keeps plot-series/FFT/spectrogram orchestration over diagnostics-local value objects, `summary_builder.py` focuses on `RunAnalysis` orchestration plus public entrypoints, and `_types.py` owns diagnostics-local helper/value types such as `AccelStatistics`, speed-breakdown rows, and plot/spectrogram value objects. Boundary summary payloads live in `shared/boundaries/analysis_payload.py`, while `shared/boundaries/summary_serialization/` is the dedicated domain/app → payload package: `_contracts.py` owns structural typing protocols, `_findings.py` owns finding annotation/serialization, `_plots.py` owns plot and table serialization, and `_summary.py` owns top-level summary assembly. History-use-case callers still receive persisted-summary projection through injected adapter wiring instead of importing that projector here. The package `__init__.py` re-exports only high-level analysis entrypoints; PDF rendering types/helpers now live under `adapters/pdf/`.
-- `use_cases/history/`: run query/delete, PDF report generation, CSV/ZIP exports, and history-facing helper orchestration above persistence ports. `helpers.py` owns the local `HistoryRecord` TypedDict plus generic history workflow helpers, while `report_interpretation.py` owns pure report-domain interpretation consumed by the PDF mapper.
+- `use_cases/history/`: run query/delete, PDF report generation, CSV/ZIP exports, and history-facing helper orchestration above persistence ports. `helpers.py` owns the local `HistoryRecord` TypedDict plus generic history workflow helpers, `report_loader.py` owns persisted report loading/shaping, `report_cache.py` owns the in-memory PDF cache and per-key build coordination, `reports.py` is the thin coordinator service over those collaborators, and `report_interpretation.py` owns pure report-domain interpretation consumed by the PDF mapper.
 - `use_cases/run/`: recording pipeline orchestration; `lifecycle_state.py` owns `RunLifecycleState` plus the active-run snapshot, `persistence_writer.py` owns history-write coordination/retry bookkeeping, `sample_flush.py` owns sample-building/flush decisions and auto-stop checks, `status_reporting.py` owns the status/health payload helpers, `_recorder_types.py` owns `RunRecorderConfig`, `RecorderShutdownReport`, and recorder metadata/shutdown helpers, `_recorder_runtime.py` owns the periodic async run loop plus recorder-runtime normalization helpers, `logger.py` owns the thin `RunRecorder` orchestrator, `post_analysis.py` owns the background analysis queue above the injected run-persistence/analysis/error boundary, and `sample_builder.py` owns pure sample-building helpers.
 - `use_cases/updates/`: wheel-based updater workflow orchestration, runtime prerequisite validation, firmware cache, ESP flashing, release discovery, install, rollback, runner, Wi-Fi, and status tracking. `firmware_cache.py` stays the thin public cache/CLI facade, `firmware_types.py` owns firmware cache/release contracts, `firmware_bundle.py` owns bundle filesystem validation/extraction/metadata helpers, `firmware_release_fetcher.py` owns GitHub firmware HTTP access, `esp_flash_manager.py` owns the flash orchestration state machine, `esp_flash_types.py` owns ESP flash contracts/state/protocols, and `esp_serial.py` / `esp_flash_runner.py` own serial discovery and subprocess execution helpers.
 - `shared/`: cross-cutting typed payloads and ports (`shared/types/` and `shared/ports.py`), boundary serializers/decoders (`shared/boundaries/`), exceptions (`shared/exceptions.py`), JSON helpers (`shared/json_utils.py`), location identifiers (`shared/locations.py`), run-context helpers, and shared pure vehicle-order math (`shared/order_bands.py`). `shared/ports.py` owns the current run-history persistence, active-client lookup, latest-sample/latest-metrics, resolved-speed, and read-only settings ports used by runtime/history/recording flows; `shared/types/` stays focused on payload/model types such as `sensor_frame.py`, `health_snapshot.py`, the feature-scoped Pydantic HTTP contract package `shared/types/api_models/`, and the typed backend/API payloads; `shared/boundaries/run_log.py` owns JSONL run-log decoding/normalization shared by diagnostics and persistence; and `shared/boundaries/diagnostic_case.py` owns persisted-summary decoding plus summary projection, while typed speed and metadata-to-domain factories live on `SpeedProfile`, `Car`, and `Symptom`.
